fixed nested functions and made annotaion module its own py file
